Refactor shapes/curves: Removed `LastValueAccessed` everywhere
Besides LoadShape and PriceShape, the others tend to be too small to justify tracking or even using a binary search. PriceShape needs some testing but it only has one (!) example.

+++ src/Common/Utilities.pas
@@ -131,7 +131,7 @@ function NameIfNotNil(obj: TDSSObject): String;
 function FullNameIfNotNil(obj: TDSSObject): String;
 
 function LowerBound(data: PSingleArray0; npts: Integer; Stride: Integer; value: Single): Integer; overload;
-function LowerBound(data: PDoubleArray0; npts: Integer; Stride: Integer; value: Single): Integer; overload;
+function LowerBound(data: PDoubleArray0; npts: Integer; Stride: Integer; value: Double): Integer; overload;
 
 implementation
 
@@ -2431,7 +2431,7 @@ begin
     Result := l;
 end;
 
-function LowerBound(data: PDoubleArray0; npts: Integer; Stride: Integer; value: Single): Integer; overload;
+function LowerBound(data: PDoubleArray0; npts: Integer; Stride: Integer; value: Double): Integer; overload;
 var
     l, r, m: Integer;
     moffset: Int64;
